[Util] - fixed some cast issues

+++ xbmc/Util.cpp
@@ -355,7 +355,7 @@ void CUtil::GetHomePath(std::string& strPath, const std::string& strTarget)
 #if defined(TARGET_DARWIN)
     int      result = -1;
     char     given_path[2*MAXPATHLEN];
-    uint32_t path_size =2*MAXPATHLEN;
+    size_t path_size =2*MAXPATHLEN;
 
     result = CDarwinUtils::GetExecutablePath(given_path, &path_size);
     if (result == 0)
@@ -1653,7 +1653,7 @@ std::string CUtil::ResolveExecutablePath()
   delete[] buf;
 #elif defined(TARGET_DARWIN)
   char     given_path[2*MAXPATHLEN];
-  uint32_t path_size =2*MAXPATHLEN;
+  size_t path_size =2*MAXPATHLEN;
 
   CDarwinUtils::GetExecutablePath(given_path, &path_size);
   strExecutablePath = given_path;
@@ -1696,7 +1696,7 @@ std::string CUtil::GetFrameworksPath(bool forPython)
   std::string strFrameworksPath;
 #if defined(TARGET_DARWIN)
   char     given_path[2*MAXPATHLEN];
-  uint32_t path_size =2*MAXPATHLEN;
+  size_t path_size =2*MAXPATHLEN;
 
   CDarwinUtils::GetFrameworkPath(forPython, given_path, &path_size);
   strFrameworksPath = given_path;
